Komal/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
openai
glob2
pandas
mysql-connector-python
